refactoring login page to be first
