--- webpage/login.html
@@ -1,31 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Login</title>
-    <link rel = "stylesheet" href="styles/login.css">
-
-    <script>var exports = {};</script>
-    <script defer src="../module_registry/out/user/user.js" type="module"></script>
-    <script defer src="scripts/io_ops.js" type="module"></script>
-    <script defer src="scripts/loginScript.js" type="module"></script>
-
-</head>
-<body>
-<div class="Box">
-    <h2>Login</h2>
-    <div id = "error"></div>
-    <form id ="form" action="/" method="post">
-        <label for="username">Username:</label>
-        <input type="text" id="username" name="username" required>
-        <label for="password">Password:</label>
-        <input type="password" id="password" name="password" required>
-        <input type="submit" value="Login">
-    </form>
-</div>
-</body>
-
-
-</html>
